move server config out of build config
